Add pivot index method to Solution class
Implement pivot index calculation in the Solution class. First calculate the total sum of the entire array. Then assume the leftSum as 0 at the beginning, then update it if it not matches the criteria. RightSum is calculated by subtracting the total sum and left sum and the current element (because we don't want to add the pivot index element, it gives inaccuracy).